ov102: rename BobOmb to daBmb_c and promote the 35-function TU
The ROM names this class. ov102 0x0214e4fc holds the bytes "7daBmb_c"
NUL-terminated -- the length-prefixed Itanium _ZTS payload -- _ZTI7daBmb_c
at 0x0214e508 points its +4 word at it, and the vtable's -4 header word
points back at the _ZTI. "BobOmb" appears nowhere in the cartridge: the
scout scanned 106 images / 3,582,144 bytes and got 0 hits. Of 121 promoted
TUs, 111 already use a ROM-style d*_c name.

The rename and the promotion are ONE commit because dsd resolves .data
records by NAME: a half-landed rename links the vtable word the factory
stores as 0x00000000.

  include/BobOmb.h -> include/daBmb_c.h, struct BobOmb -> struct daBmb_c
  config/arm9/overlays/ov102/symbols.txt: nine member rows _ZN6BobOmb* ->
    _ZN7daBmb_c* (only the class prefix moves), and the duplicate coined
    _ZTV6BobOmb alias row at 0x0214e558 is dropped -- exactly one spelling
    can be what the data records resolve through
  include/decl_common.h: the _ZTV declaration
  symbols/actor_renames.tsv: fourteen appended rows

PROMOTION. The 35 shards of the contiguous run 0x0214a96c..0x0214c748
collapse into src/actors/daBmb_c.cpp.

  35 shards deleted; delinks.txt 35 blocks -> 1 (CRLF preserved)
  config/tu_manifest.d/ov102/daBmb_c.json: status promoted, .text claim
    0x0214a96c..0x0214c748, 35 functions, 12 compiler_only_output rows
  tubuild.py verify   : 35/35 MATCH, objisolate clean, reloc destinations
                        clean, emission order ROM-ascending -> TEXT-VERIFIED
  tubuild.py linkcheck: SCRATCH-LINK-VERIFIED before promotion; after it,
                        106/106 modules exact, 100.000000% of compared
                        bytes, ROM sha256 d1506e90...c478e8, identical to
                        the stock control

The run is 35, not the 34 tu_map reports: tu_map segments on symbol NAME,
so daBmb_c_classInit (0x0214c6f8, size 0x50) falls outside a run it
physically abuts. Absorbing it needed the vptr store to become
(int)(_ZTV7daBmb_c + 2): this TU now DEFINES the vtable, because with the
destructor inline in the header OnYoshiTryEat is the first out-of-line
virtual the class declares and therefore the key function. The bare
spelling the shard used would bind to that local definition and store a
vptr eight bytes low; every byte gate wildcards the relocated word, and
only the link sees it.

Two members needed measurement to match inside one TU:

  func_ov102_0214b53c -- the shard's chain of `& 0xFFFFFFFFu` masks is a
    C-ONLY lever. Under -lang c++ the `&&` is a bool, the mask forces
    mwccarm to materialise it in a register, and the function grows three
    instructions. Plain short-circuit restores the cartridge's branches.
  func_ov102_0214baa0 -- VOID, though OnTurnIntoEgg's shard declared it int
    and forwarded its r0. With a non-void return type mwccarm keeps r0
    reserved and the final store materialises its zero in r1. Void restores
    r0; the call site pays nothing, because `bl` + the caller's own epilogue
    is what both spellings emit.

STAGE 3b. Thirteen of the 35 are C++ methods: the nine the cartridge names,
plus four of the six mState arms under coined State<N> names -- the index is
ROM-proven (func_ov102_0214b03c switches on mState and dispatches one body
per value 0..5) and the name claims the index and nothing more. 0x390 is now
mCarrier, a Player *: func_ov102_0214b53c passes it as the Player& argument
of dActor_c::UpdateCarry and func_ov102_0214ae1c calls Player::DropActor on
it. notes/enemy-leaf-provenance.md had it listed as never read anywhere in
the tree; that was true only of the five shards that section had read.

The other 21 stay free functions, and the wall is never codegen:

  SCOPE (2) -- mState arms 0 and 2 compile to the ROM's exact sizes as
    members and the TU then will not LINK. As members they lose their own
    block-scope extern "C" declarations, and the one file-scope region
    cannot take them: other members already own contradicting spellings,
    decisively func_0200fc44, which arm 2 recovered with four parameters and
    func_ov102_0214aa18 with three. Both byte-match their own body.
    tubuild.py verify called the broken form 35/35 MATCH with reloc
    destinations clean -- only mwldarm saw it.
  NAMING (19) -- the cartridge carries their addresses and no identifier,
    and three of them (0x0214ad14, 0x0214ae1c, 0x0214b384) are called from
    ov002/ov014/ov078/ov098, so naming those is a cross-module rename.

CONVERTED ratchet: 2696 -> 2696. The six banked shard identities become six
member identities of the merged file; the removals are logged with the usual
reason, that a reconstructed TU must spell vague-linkage symbols directly or
the range will not link.

  layout_check            : clean
  check_src_tu            : 30 TUs, 783 mangled refs, every reference resolves
  check_tubuild_conflicts : 154 entries, every recorded conflict matches
  check_dead_references   : only the pre-existing dScMgSound_c.cpp reference
  cpp_tu_compat --require-ready : 9/9 READY
  test_tu_manifest        : 25 tests OK

notes/cpp-tu-current-state.md regenerated. The queue row is re-derived to
`daBmb_c 1 1636 ov102 yes yes`; queue_audit's other disagreement, a
dScMgLuigi_c line count, is pre-existing on main and was left alone.

class_rename.py was not run; every edit here is an ordinary one.
