+++ config/arm9/overlays/ov102/delinks.txt
@@ -216,145 +216,9 @@ src/d_a_obj_hatena_block_vs_item_block.c:
     complete
     .text start:0x0214a92c end:0x0214a96c
 
-src/_ZN6BobOmbD1Ev.cpp:
+src/actors/daBmb_c.cpp:
     complete
-    .text start:0x0214a96c end:0x0214a9b4
-
-src/_ZN6BobOmbD0Ev.cpp:
-    complete
-    .text start:0x0214a9b4 end:0x0214aa10
-
-src/_ZN6BobOmb16OnAimedAtWithEggEv.cpp:
-    complete
-    .text start:0x0214aa10 end:0x0214aa18
-
-src/func_ov102_0214aa18.cpp:
-    complete
-    .text start:0x0214aa18 end:0x0214ab1c
-
-src/func_ov102_0214ab1c.c:
-    complete
-    .text start:0x0214ab1c end:0x0214ad14
-
-src/func_ov102_0214ad14.c:
-    complete
-    .text start:0x0214ad14 end:0x0214ad40
-
-src/func_ov102_0214ad40.c:
-    complete
-    .text start:0x0214ad40 end:0x0214adc8
-
-src/_ZN6BobOmb13OnTurnIntoEggER6Player.cpp:
-    complete
-    .text start:0x0214adc8 end:0x0214ae1c
-
-src/func_ov102_0214ae1c.c:
-    complete
-    .text start:0x0214ae1c end:0x0214b03c
-
-src/func_ov102_0214b03c.cpp:
-    complete
-    .text start:0x0214b03c end:0x0214b128
-
-src/func_ov102_0214b128.cpp:
-    complete
-    .text start:0x0214b128 end:0x0214b248
-
-src/func_ov102_0214b248.cpp:
-    complete
-    .text start:0x0214b248 end:0x0214b384
-
-src/func_ov102_0214b384.c:
-    complete
-    .text start:0x0214b384 end:0x0214b3b8
-
-src/func_ov102_0214b3b8.c:
-    complete
-    .text start:0x0214b3b8 end:0x0214b3f0
-
-src/func_ov102_0214b3f0.c:
-    complete
-    .text start:0x0214b3f0 end:0x0214b444
-
-src/func_ov102_0214b444.cpp:
-    complete
-    .text start:0x0214b444 end:0x0214b53c
-
-src/func_ov102_0214b53c.c:
-    complete
-    .text start:0x0214b53c end:0x0214b988
-
-src/func_ov102_0214b988.cpp:
-    complete
-    .text start:0x0214b988 end:0x0214ba30
-
-src/func_ov102_0214ba30.cpp:
-    complete
-    .text start:0x0214ba30 end:0x0214baa0
-
-src/func_ov102_0214baa0.c:
-    complete
-    .text start:0x0214baa0 end:0x0214bbd8
-
-src/func_ov102_0214bbd8.c:
-    complete
-    .text start:0x0214bbd8 end:0x0214bc20
-
-src/func_ov102_0214bc20.c:
-    complete
-    .text start:0x0214bc20 end:0x0214bcc8
-
-src/func_ov102_0214bcc8.c:
-    complete
-    .text start:0x0214bcc8 end:0x0214bd20
-
-src/func_ov102_0214bd20.cpp:
-    complete
-    .text start:0x0214bd20 end:0x0214bd90
-
-src/func_ov102_0214bd90.c:
-    complete
-    .text start:0x0214bd90 end:0x0214be1c
-
-src/func_ov102_0214be1c.cpp:
-    complete
-    .text start:0x0214be1c end:0x0214beb4
-
-src/func_ov102_0214beb4.c:
-    complete
-    .text start:0x0214beb4 end:0x0214bf64
-
-src/func_ov102_0214bf64.cpp:
-    complete
-    .text start:0x0214bf64 end:0x0214c0b8
-
-src/func_ov102_0214c0b8.c:
-    complete
-    .text start:0x0214c0b8 end:0x0214c12c
-
-src/_ZN6BobOmb16CleanupResourcesEv.cpp:
-    complete
-    .text start:0x0214c12c end:0x0214c168
-
-src/_ZN6BobOmb6RenderEv.cpp:
-    complete
-    .text start:0x0214c168 end:0x0214c1bc
-
-src/_ZN6BobOmb8BehaviorEv.cpp:
-    complete
-    .text start:0x0214c1bc end:0x0214c510
-
-src/_ZN6BobOmb13InitResourcesEv.cpp:
-    complete
-    .text start:0x0214c510 end:0x0214c6e4
-
-src/_ZN6BobOmb13OnYoshiTryEatEv.cpp:
-    complete
-    .text start:0x0214c6e4 end:0x0214c6f8
-
-src/d_a_bmb.c:
-    complete
-    .text start:0x0214c6f8 end:0x0214c748
+    .text start:0x0214a96c end:0x0214c748
 
 src/_ZN10KoopaShellD1Ev.cpp:
     complete
