+++ config/arm9/overlays/ov102/symbols.txt
@@ -51,14 +51,14 @@ daObjHatenaBlock_c_classInit_CAP_BLOCK_M kind:function(arm,size=0x40) addr:0x021
 daObjHatenaBlock_c_classInit_ITEM_BLOCK kind:function(arm,size=0x40) addr:0x0214a8ac
 daObjHatenaBlock_c_classInit_HATENA_BLOCK kind:function(arm,size=0x40) addr:0x0214a8ec
 daObjHatenaBlock_c_classInit_VS_ITEM_BLOCK kind:function(arm,size=0x40) addr:0x0214a92c
-_ZN6BobOmbD1Ev kind:function(arm,size=0x48) addr:0x0214a96c
-_ZN6BobOmbD0Ev kind:function(arm,size=0x5c) addr:0x0214a9b4
-_ZN6BobOmb16OnAimedAtWithEggEv kind:function(arm,size=0x8) addr:0x0214aa10
+_ZN7daBmb_cD1Ev kind:function(arm,size=0x48) addr:0x0214a96c
+_ZN7daBmb_cD0Ev kind:function(arm,size=0x5c) addr:0x0214a9b4
+_ZN7daBmb_c16OnAimedAtWithEggEv kind:function(arm,size=0x8) addr:0x0214aa10
 func_ov102_0214aa18 kind:function(arm,size=0x104) addr:0x0214aa18
 func_ov102_0214ab1c kind:function(arm,size=0x1f8) addr:0x0214ab1c
 func_ov102_0214ad14 kind:function(arm,size=0x2c) addr:0x0214ad14
 func_ov102_0214ad40 kind:function(arm,size=0x88) addr:0x0214ad40
-_ZN6BobOmb13OnTurnIntoEggER6Player kind:function(arm,size=0x54) addr:0x0214adc8
+_ZN7daBmb_c13OnTurnIntoEggER6Player kind:function(arm,size=0x54) addr:0x0214adc8
 func_ov102_0214ae1c kind:function(arm,size=0x220) addr:0x0214ae1c
 func_ov102_0214b03c kind:function(arm,size=0xec) addr:0x0214b03c
 func_ov102_0214b128 kind:function(arm,size=0x120) addr:0x0214b128
@@ -69,22 +69,22 @@ func_ov102_0214b3f0 kind:function(arm,size=0x54) addr:0x0214b3f0
 func_ov102_0214b444 kind:function(arm,size=0xf8) addr:0x0214b444
 func_ov102_0214b53c kind:function(arm,size=0x44c) addr:0x0214b53c
 func_ov102_0214b988 kind:function(arm,size=0xa8) addr:0x0214b988
-func_ov102_0214ba30 kind:function(arm,size=0x70) addr:0x0214ba30
+_ZN7daBmb_c6State5Ev kind:function(arm,size=0x70) addr:0x0214ba30
 func_ov102_0214baa0 kind:function(arm,size=0x138) addr:0x0214baa0
-func_ov102_0214bbd8 kind:function(arm,size=0x48) addr:0x0214bbd8
+_ZN7daBmb_c6State4Ev kind:function(arm,size=0x48) addr:0x0214bbd8
 func_ov102_0214bc20 kind:function(arm,size=0xa8) addr:0x0214bc20
-func_ov102_0214bcc8 kind:function(arm,size=0x58) addr:0x0214bcc8
+_ZN7daBmb_c6State3Ev kind:function(arm,size=0x58) addr:0x0214bcc8
 func_ov102_0214bd20 kind:function(arm,size=0x70) addr:0x0214bd20
 func_ov102_0214bd90 kind:function(arm,size=0x8c) addr:0x0214bd90
-func_ov102_0214be1c kind:function(arm,size=0x98) addr:0x0214be1c
+_ZN7daBmb_c6State1Ev kind:function(arm,size=0x98) addr:0x0214be1c
 func_ov102_0214beb4 kind:function(arm,size=0xb0) addr:0x0214beb4
 func_ov102_0214bf64 kind:function(arm,size=0x154) addr:0x0214bf64
 func_ov102_0214c0b8 kind:function(arm,size=0x74) addr:0x0214c0b8
-_ZN6BobOmb16CleanupResourcesEv kind:function(arm,size=0x3c) addr:0x0214c12c
-_ZN6BobOmb6RenderEv kind:function(arm,size=0x54) addr:0x0214c168
-_ZN6BobOmb8BehaviorEv kind:function(arm,size=0x354) addr:0x0214c1bc
-_ZN6BobOmb13InitResourcesEv kind:function(arm,size=0x1d4) addr:0x0214c510
-_ZN6BobOmb13OnYoshiTryEatEv kind:function(arm,size=0x14) addr:0x0214c6e4
+_ZN7daBmb_c16CleanupResourcesEv kind:function(arm,size=0x3c) addr:0x0214c12c
+_ZN7daBmb_c6RenderEv kind:function(arm,size=0x54) addr:0x0214c168
+_ZN7daBmb_c8BehaviorEv kind:function(arm,size=0x354) addr:0x0214c1bc
+_ZN7daBmb_c13InitResourcesEv kind:function(arm,size=0x1d4) addr:0x0214c510
+_ZN7daBmb_c13OnYoshiTryEatEv kind:function(arm,size=0x14) addr:0x0214c6e4
 daBmb_c_classInit kind:function(arm,size=0x50) addr:0x0214c6f8
 _ZN10KoopaShellD1Ev kind:function(arm,size=0x50) addr:0x0214c748
 _ZN10KoopaShellD0Ev kind:function(arm,size=0x64) addr:0x0214c798
@@ -188,7 +188,6 @@ _ZTI7daBmb_c kind:data(any) addr:0x0214e508
 data_ov102_0214e514 kind:data(any) addr:0x0214e514
 data_ov102_0214e524 kind:data(any) addr:0x0214e524
 g_profile_BOMBHEI kind:data(any) addr:0x0214e534
-_ZTV6BobOmb kind:data(any) addr:0x0214e558
 _ZTV7daBmb_c kind:data(any) addr:0x0214e558
 data_ov102_0214e5d4 kind:data(any) addr:0x0214e5d4
 data_ov102_0214e5dc kind:data(any) addr:0x0214e5dc
