+++ config/converted-baseline.json
@@ -1305,12 +1305,6 @@
     "src/_ZN5dPa_c7level_c20edStarKiraCallback_cC1Ev.cpp",
     "src/_ZN5dPa_c7level_c23cleanParticleCallback_c8OnUpdateERN8Particle6SystemEb.cpp",
     "src/_ZN5dPa_c7level_c24fitWaterSimpleCallback_c8OnUpdateERN8Particle6SystemEb.cpp",
-    "src/_ZN6BobOmb13OnYoshiTryEatEv.cpp",
-    "src/_ZN6BobOmb16CleanupResourcesEv.cpp",
-    "src/_ZN6BobOmb16OnAimedAtWithEggEv.cpp",
-    "src/_ZN6BobOmb6RenderEv.cpp",
-    "src/_ZN6BobOmbD0Ev.cpp",
-    "src/_ZN6BobOmbD1Ev.cpp",
     "src/_ZN6Bowser16CleanupResourcesEv.cpp",
     "src/_ZN6Bowser16OnPendingDestroyEv.cpp",
     "src/_ZN6Bowser8BehaviorEv.cpp",
@@ -2156,6 +2150,12 @@
     "src/actors/dScMgSound_c.cpp#_ZN12dScMgSound_cD1Ev",
     "src/actors/dScMgTeresa_c.cpp#_ZN13dScMgTeresa_cD0Ev",
     "src/actors/dScMgTeresa_c.cpp#_ZN13dScMgTeresa_cD1Ev",
+    "src/actors/daBmb_c.cpp#_ZN7daBmb_c13OnYoshiTryEatEv",
+    "src/actors/daBmb_c.cpp#_ZN7daBmb_c16CleanupResourcesEv",
+    "src/actors/daBmb_c.cpp#_ZN7daBmb_c16OnAimedAtWithEggEv",
+    "src/actors/daBmb_c.cpp#_ZN7daBmb_c6RenderEv",
+    "src/actors/daBmb_c.cpp#_ZN7daBmb_cD0Ev",
+    "src/actors/daBmb_c.cpp#_ZN7daBmb_cD1Ev",
     "src/actors/daDgr_c.cpp#_ZN7daDgr_c16CleanupResourcesEv",
     "src/actors/daDgr_c.cpp#_ZN7daDgr_c6RenderEv",
     "src/actors/daDgr_c.cpp#_ZN7daDgr_cD0Ev",
