+++ include/daBmb_c.h
@@ -1,9 +1,9 @@
-#ifndef BOBOMB_H
-#define BOBOMB_H
+#ifndef DABMB_C_H
+#define DABMB_C_H
 
 #include "types.h"
 
-/* Derives from dEnemyBase_c, on the evidence of its own destructor: `_ZN6BobOmbD1Ev`
+/* Derives from dEnemyBase_c, on the evidence of its own destructor: `_ZN7daBmb_cD1Ev`
  * stores this vtable, destroys its members in reverse declaration order, then
  * calls `dEnemyBase_c::~dEnemyBase_c`. Everything this header used to restate below 0x110
  * belongs to that chain and is inherited now.
@@ -24,14 +24,18 @@
  * also rename things its callers spell.
  *
  * SIZE IS THE ROM'S OWN, not a rounded-up field span: `daBmb_c_classInit` calls
- * `fBase_c::operator new(1024)` -- 0x400 -- and stores `_ZTV6BobOmb`,
+ * `fBase_c::operator new(1024)` -- 0x400 -- and stores `_ZTV7daBmb_c`,
  * so that literal IS this class's sizeof. The observed fields only span to
  * 0x3f8; the difference is trailing space no source reads.
  *
- * SM64DS RTTI names the implementation daBmb_c. The reconstructed
- * factory daBmb_c_classInit (historical alias
- * BobOmb_Spawn) constructs it for the BOMBHEI
- * registry profile.
+ * THE NAME IS THE CARTRIDGE'S OWN. ov102 0x0214e4fc holds the bytes
+ * "7daBmb_c\0" -- the length-prefixed mangled type name -- and _ZTI7daBmb_c at
+ * 0x0214e508 points its +4 word back at that string, while the vtable's -4
+ * header word points at the _ZTI. The class was carried here under the coined
+ * name BobOmb, which occurs nowhere in the cartridge; that spelling is gone and
+ * every member now mangles as _ZN7daBmb_c*. The reconstructed factory
+ * daBmb_c_classInit (historical alias BobOmb_Spawn) constructs it for the
+ * BOMBHEI registry profile.
  */
 
 #include "dEnemyBase_c.h"
@@ -43,7 +47,7 @@
 #include "TextureTransformer.h"
 #include "dBgCh_Actr.h"
 
-struct BobOmb : dEnemyBase_c {
+struct daBmb_c : dEnemyBase_c {
     dCcAc_c           mdCc_c;         /* 0x110 */
     dBgCh_Actr                 mWithMeshClsn;         /* 0x144 */
     ModelAnim                    mModelAnim;            /* 0x300 */
@@ -52,9 +56,16 @@ struct BobOmb : dEnemyBase_c {
     /* Everything from here down was reachable only through the `struct Obj`
        shadow InitResources used to carry, so the generated header never had
        it. mMatrix is a 0x30-byte block copied wholesale from IDENTITY_MATRIX4X3. */
-    u32                          unk_390;               /* 0x390 -- InitResources zeroes it; no reader */
+    /* The Player carrying this Bob-omb, or 0.  Proven by use, not guessed:
+       func_ov102_0214b53c passes it as the `Player &` argument of
+       dActor_c::UpdateCarry(Player &, Vector3 const &) and calls
+       Player::IsFrontSliding and Player::LostGrabbedObject on it;
+       func_ov102_0214ae1c calls Player::DropActor and Player::Hurt on it and
+       then clears it; func_ov102_0214b3f0 is the setter and
+       func_ov102_0214b3b8 the clearer.  InitResources zeroes it. */
+    Player                      *mCarrier;              /* 0x390 */
     /* A Matrix4x3, copied wholesale from IDENTITY_MATRIX4X3 by InitResources.
-       Left as twelve words rather than typed: BobOmb.h is included by files that
+       Left as twelve words rather than typed: daBmb_c.h is included by files that
        do not pull common.h, and embedding the real type would force it on them. */
     s32                          mMatrix[12];           /* 0x394 */
     s32                          mHomePosX;             /* 0x3c4 */
@@ -88,9 +99,21 @@ struct BobOmb : dEnemyBase_c {
     u8  pad_3f7[0x9];
 
     /* --- vtable --- */
-    virtual ~BobOmb();
 
-    virtual s32   OnYoshiTryEat();         /* slot 18 */
+    /* INLINE, AND DECLARED FIRST. The cartridge puts D1 at 0x0214a96c below D0
+       at 0x0214a9b4 and carries no D2 anywhere, which is what mwccarm 2004/b56
+       emits for an inline in-class destructor; the out-of-line form emits
+       D2/D0/D1 in the wrong order plus a homeless D2. The typed member list
+       above makes the empty body own the ShadowModel, ModelAnim, dBgCh_Actr and
+       dCcAc_c teardowns and the chain into _ZN12dEnemyBase_cD2Ev.
+
+       With the destructor inline, OnYoshiTryEat becomes the first out-of-line
+       virtual this class declares -- the key function -- so the vtable and the
+       RTTI group land in the translation unit that defines it,
+       src/actors/daBmb_c.cpp. */
+    virtual ~daBmb_c() {}
+
+    virtual s32   OnYoshiTryEat();         /* slot 18 -- key function */
     virtual int   OnTurnIntoEgg(Player &player); /* slot 19 */
     virtual s32   OnAimedAtWithEgg();      /* slot 29 */
 
@@ -98,8 +121,22 @@ struct BobOmb : dEnemyBase_c {
     int CleanupResources();
     int InitResources();
     int Render();
+
+    /* Four of the six arms of the mState state machine.  The NUMBER is what the
+       cartridge proves: func_ov102_0214b03c switches on mState (+0x3dc) and
+       dispatches exactly six bodies for 0..5, one each, and nothing else reads the
+       field to choose between them.  The NAMES are coined -- ov102 carries these
+       addresses and no identifier -- so they claim the index and nothing more; what
+       each arm does is written at its body.  Arms 0 and 2 stay free functions
+       (func_ov102_0214bf64, func_ov102_0214bd90): their recovered bodies carry
+       block-scope extern "C" declarations that contradict declarations other members
+       of this TU own, and a member function may not hold a linkage specification. */
+    void State1();
+    void State3();
+    int  State4();
+    void State5();
 };
 
-typedef char BobOmb_size_must_be_0x400[sizeof(BobOmb) == 0x400 ? 1 : -1];
+typedef char daBmb_c_size_must_be_0x400[sizeof(daBmb_c) == 0x400 ? 1 : -1];
 
-#endif /* BOBOMB_H */
+#endif /* DABMB_C_H */
