+++ include/decl_common.h
@@ -701,7 +701,7 @@ extern int _ZTV6Fwoosh[];
 extern int _ZTV5Swoop[];
 extern int _ZTV5Unagi[];
 extern int _ZTV5Whomp[];
-extern int _ZTV6BobOmb[];
+extern int _ZTV7daBmb_c[];
 extern int _ZTV6Bowser[];
 extern int _ZTV6Bullet[];
 extern int _ZTV6Cannon[];
@@ -2906,7 +2906,6 @@ extern void func_ov102_0214ad14(void*);
 extern void func_ov102_0214ae1c(void*);
 extern void func_ov102_0214b03c(void*);
 extern void func_ov102_0214baa0(void*);
-extern void func_ov102_0214be1c(char*);
 extern void func_ov102_0214c7fc(char*);
 extern void func_ov102_0214c84c(char*);
 extern void func_ov102_0214cbec(char*);
