--- src/_ZN6BobOmb13OnYoshiTryEatEv.cpp
@@ -1,8 +1,0 @@
-//cpp
-// @symbol _ZN6BobOmb13OnYoshiTryEatEv
-#include "BobOmb.h"
-/* recovered: renamed to Class_Method */
-s32 BobOmb::OnYoshiTryEat() {
-    /* 263 == 0x107, dEnemyBase_c's own byte at that offset. */
-    return mEatenByYoshi == 0;
-}
