--- src/_ZN6BobOmb16OnAimedAtWithEggEv.cpp
@@ -1,9 +1,0 @@
-//cpp
-// @symbol _ZN6BobOmb16OnAimedAtWithEggEv
-#include "BobOmb.h"
-// recovered name: BobOmb_OnAimedAtWithEgg
-/* recovered: renamed to Class_Method */
-/* daBmb_c::OnAimedAtWithEgg - recovered from vtable slot identity */
-s32 BobOmb::OnAimedAtWithEgg() {
-    return 204800;
-}
