--- src/_ZN6BobOmbD0Ev.cpp
@@ -1,16 +1,0 @@
-//cpp
-// @symbol _ZN6BobOmbD0Ev
-/* recovered: real C++ deleting destructor -- the compiler emits the whole body
- *
- * D0 is the DELETING destructor: destroy through this class and its bases, then
- * return the object to its heap. Declaring `~BobOmb()` is enough -- mwcc emits
- * D2, D0 and D1 together and objisolate keeps the one this file is bound to.
- *
- * The deallocation is an inline operator delete -- dEnemyBase_c's, reachable because
- * dEnemyBase_c is this class's IMMEDIATE base.
- */
-#include "BobOmb.h"
-
-BobOmb::~BobOmb()
-{
-}
