--- src/func_ov102_0214b384.c
@@ -1,8 +1,0 @@
-void func_ov102_0214b384(void *arg0, unsigned int arg1) {
-    char *p = (char *)arg0;
-    unsigned int cur = *(unsigned short *)(p + 0x3ea);
-    if (cur == 0 || cur > arg1) {
-        *(unsigned short *)(p + 0x3ea) = (unsigned short)arg1;
-    }
-    *(unsigned int *)(p + 0xb0) &= ~1u;
-}
